🚸 added url access for each component in documentation
