+++ doc/src/components/WithComponentsBar/styles.scss
